requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pyodbc
SQLAlchemy

# note had to brew install unixodbc for pyodbc to work
# also had to brew install msodbcsql mssql-tools